Test infra for py2.7 is broken. Drop it.
We will no longer test packs with python2.7, as it is broken and the fix
(re-adding 2.7) is too invasive across all of our infrastructure.

We continue supporting 2.7 but only on an as-reported basis.
If someone submits a PR for 2.7 and it doesn't break with python 3, great.
If a pack is upgraded so that it can only be used with python 3,
then people can downgrade to an earlier version that supported python 2.

https://github.com/StackStorm/discussions/issues/40#issuecomment-769992629